* extend tests * remove start_all
