+++ command/app/apply_test.go
@@ -44,6 +44,11 @@ func TestCmdApplyWithErrorForApplicationService(t *testing.T) {
 
 	fakeClientSet := new(fake.Clientset)
 
+	oldImagesLoader := imagesService
+	imagesLoaderMock := new(_mocks.ImagesInterface)
+
+	imagesService = imagesLoaderMock
+
 	serviceBuilderMock.On("GetClientSet", config).Return(fakeClientSet, nil)
 	serviceBuilderMock.On("GetApplicationService", fakeClientSet, "foobar", config).Return(nil, errors.New("explode"))
 
@@ -51,6 +56,7 @@ func TestCmdApplyWithErrorForApplicationService(t *testing.T) {
 		cli.OsExiter = oldHandler
 		configLoader = oldConfigLoader
 		serviceBuilder = oldServiceBuilder
+		imagesService = oldImagesLoader
 	}()
 
 	cli.OsExiter = func(exitCode int) {
@@ -61,6 +67,110 @@ func TestCmdApplyWithErrorForApplicationService(t *testing.T) {
 	})
 
 	assert.Equal(t, "explode\n", errOutput)
+	assert.Empty(t, output)
+}
+
+func TestCmdApplyWithErrorForHasTag(t *testing.T) {
+
+	oldHandler := cli.OsExiter
+
+	oldConfigLoader := configLoader
+	configLoaderMock := new(_mocks.ConfigLoaderInterface)
+
+	configLoader = configLoaderMock
+
+	config := loader.Config{
+		ProjectID: "test-project",
+		Zone:      "berlin",
+		ClusterID: "testing",
+	}
+
+	configLoaderMock.On("LoadConfigFromPath", "never.yml").Return(config, nil)
+
+	oldServiceBuilder := serviceBuilder
+	serviceBuilderMock := new(_mocks.BuilderInterface)
+
+	serviceBuilder = serviceBuilderMock
+
+	fakeClientSet := new(fake.Clientset)
+	fakeApplicationService := new(_mocks.ApplicationServiceInterface)
+	oldImagesLoader := imagesService
+	imagesLoaderMock := new(_mocks.ImagesInterface)
+
+	imagesService = imagesLoaderMock
+
+	imagesLoaderMock.On("HasTag", config.Cleanup, "staging-latest").Return(false, errors.New("explode"))
+
+	serviceBuilderMock.On("GetClientSet", config).Return(fakeClientSet, nil)
+	serviceBuilderMock.On("GetApplicationService", fakeClientSet, "staging", config).Return(fakeApplicationService, nil)
+
+	defer func() {
+		cli.OsExiter = oldHandler
+		configLoader = oldConfigLoader
+		serviceBuilder = oldServiceBuilder
+		imagesService = oldImagesLoader
+	}()
+
+	cli.OsExiter = func(exitCode int) {
+		assert.Equal(t, 1, exitCode)
+	}
+	output, errOutput := captureOutput(func() {
+		command.RunTestCommand(CmdApply, []string{"apply", "-c", "never.yml", "master"})
+	})
+
+	assert.Equal(t, "explode\n", errOutput)
+	assert.Empty(t, output)
+}
+
+func TestCmdApplyWithFalseForHasTag(t *testing.T) {
+
+	oldHandler := cli.OsExiter
+
+	oldConfigLoader := configLoader
+	configLoaderMock := new(_mocks.ConfigLoaderInterface)
+
+	configLoader = configLoaderMock
+
+	config := loader.Config{
+		ProjectID: "test-project",
+		Zone:      "berlin",
+		ClusterID: "testing",
+	}
+
+	configLoaderMock.On("LoadConfigFromPath", "never.yml").Return(config, nil)
+
+	oldServiceBuilder := serviceBuilder
+	serviceBuilderMock := new(_mocks.BuilderInterface)
+
+	serviceBuilder = serviceBuilderMock
+
+	fakeClientSet := new(fake.Clientset)
+	fakeApplicationService := new(_mocks.ApplicationServiceInterface)
+	oldImagesLoader := imagesService
+	imagesLoaderMock := new(_mocks.ImagesInterface)
+
+	imagesService = imagesLoaderMock
+
+	imagesLoaderMock.On("HasTag", config.Cleanup, "latest").Return(false, nil)
+
+	serviceBuilderMock.On("GetClientSet", config).Return(fakeClientSet, nil)
+	serviceBuilderMock.On("GetApplicationService", fakeClientSet, "production", config).Return(fakeApplicationService, nil)
+
+	defer func() {
+		cli.OsExiter = oldHandler
+		configLoader = oldConfigLoader
+		serviceBuilder = oldServiceBuilder
+		imagesService = oldImagesLoader
+	}()
+
+	cli.OsExiter = func(exitCode int) {
+		assert.Equal(t, 0, exitCode)
+	}
+	output, errOutput := captureOutput(func() {
+		command.RunTestCommand(CmdApply, []string{"apply", "-c", "never.yml", "-p"})
+	})
+
+	assert.Equal(t, "No Image 'latest' found for namespace 'production' \n", errOutput)
 	assert.Empty(t, output)
 }
 
@@ -88,6 +198,12 @@ func TestCmdApplyWithErrorForCreateApplication(t *testing.T) {
 
 	fakeClientSet := new(fake.Clientset)
 	fakeApplicationService := new(_mocks.ApplicationServiceInterface)
+	oldImagesLoader := imagesService
+	imagesLoaderMock := new(_mocks.ImagesInterface)
+
+	imagesService = imagesLoaderMock
+
+	imagesLoaderMock.On("HasTag", config.Cleanup, "staging-foobar-latest").Return(true, nil)
 
 	serviceBuilderMock.On("GetClientSet", config).Return(fakeClientSet, nil)
 	serviceBuilderMock.On("GetApplicationService", fakeClientSet, "foobar", config).Return(fakeApplicationService, nil)
@@ -98,6 +214,7 @@ func TestCmdApplyWithErrorForCreateApplication(t *testing.T) {
 		cli.OsExiter = oldHandler
 		configLoader = oldConfigLoader
 		serviceBuilder = oldServiceBuilder
+		imagesService = oldImagesLoader
 	}()
 
 	cli.OsExiter = func(exitCode int) {
@@ -136,6 +253,13 @@ func TestCmdApply(t *testing.T) {
 	fakeClientSet := new(fake.Clientset)
 	fakeApplicationService := new(_mocks.ApplicationServiceInterface)
 
+	oldImagesLoader := imagesService
+	imagesLoaderMock := new(_mocks.ImagesInterface)
+
+	imagesService = imagesLoaderMock
+
+	imagesLoaderMock.On("HasTag", config.Cleanup, "staging-foobar-latest").Return(true, nil)
+
 	serviceBuilderMock.On("GetClientSet", config).Return(fakeClientSet, nil)
 	serviceBuilderMock.On("GetApplicationService", fakeClientSet, "foobar", config).Return(fakeApplicationService, nil)
 
@@ -145,6 +269,7 @@ func TestCmdApply(t *testing.T) {
 		cli.OsExiter = oldHandler
 		configLoader = oldConfigLoader
 		serviceBuilder = oldServiceBuilder
+		imagesService = oldImagesLoader
 	}()
 
 	cli.OsExiter = func(exitCode int) {
